requirements.txt
flask
python-dotenv
requests
langchain
langchain-community

langchain-ollama
